+++ today/src/components/ChatList.js
@@ -1,3 +1,5 @@
+import "../App";
+
 import { useParams, Link } from "react-router-dom";
 
 const ChatList = (props)=>{
@@ -7,7 +9,7 @@ const ChatList = (props)=>{
     return (
     <div>
         {Object.keys(chats).map((id, index)=>(
-            <div key={index}>
+            <div key={index} className="chatItem">
                 <Link to={`/chats/${id}`}><b style={{color: id===chatId?"#88affc":"white"}}>{chats[id].name}</b>
                 </Link>
             </div>
