+++ today/src/components/MessageList.js
@@ -1,4 +1,4 @@
-import { List, Avatar, ListItem, ListItemAvatar, ListItemText } from "@mui/material";
+import { List, ListItem, ListItemIcon, ListItemText } from "@mui/material";
 import AndroidIcon from '@mui/icons-material/Android';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
 
@@ -6,13 +6,11 @@ const MessageList = (props)=>{
     const {messages} = props;
 
     return (
-        <div>
-            <List sx={{ width: '100%', maxWidth: 360, bgcolor: 'background.paper' }}>
+        <div className="messageList">
+            <List sx={{ width: '100%', maxWidth: 600}}>
                 {messages?.map((item, index)=>(
                     <ListItem key={index}>
-                        <ListItemAvatar>
-                            <Avatar>{item.user==="Bot"?<AndroidIcon/>:<AccountCircleIcon/>}</Avatar>
-                        </ListItemAvatar>
+                        <ListItemIcon >{item.user==="Bot"?<AndroidIcon/>:<AccountCircleIcon/>}</ListItemIcon>
                         <ListItemText primary={item.text} secondary={item.user} />
                     </ListItem>
                 ))}
