+++ today/src/pages/NotFound.js
@@ -2,8 +2,8 @@ import { Link } from "react-router-dom"
 
 const NotFound = ()=>{
     return <div>
-        <div>Данная страница затерялась в анналах истории, вернитесь, пожалуйста, на главную</div>
-        <Link to="/home">Тык</Link>
+        <h2>Данная страница затерялась в анналах истории, вернитесь, пожалуйста, на главную</h2>
+        <Link to="/home" className="navLink">Тык</Link>
     </div>
 }
